move IHideObjectMembers to top level namespace
