+++ Source/FakeItEasy/Creation/IFakeOptions.cs
@@ -5,7 +5,6 @@ namespace FakeItEasy.Creation
     using System.Diagnostics.CodeAnalysis;
     using System.Linq.Expressions;
     using System.Reflection.Emit;
-    using Configuration;
 
     /// <summary>
     /// Provides options for generating fake object.
